+++ packages/metascraper-helpers/test/index.js
@@ -5,6 +5,7 @@ const cheerio = require('cheerio')
 const should = require('should')
 
 const {
+  $jsonld,
   absoluteUrl,
   author,
   date,
@@ -20,7 +21,6 @@ const {
   isUrl,
   isVideoExtension,
   isVideoUrl,
-  $jsonld,
   jsonld,
   normalizeUrl,
   titleize,
@@ -233,10 +233,9 @@ describe('metascraper-helpers', () => {
     })
 
     it('reads multiple JSON-LD blocks', () => {
-      const $ = cheerio.load(`
-<script type="application/ld+json"> { "@context": "http://schema.org", "@type": "Organization", "url": "https://bykvu.com/ru", "logo": "https://bykvu.com/wp-content/themes/bykvu/img/logo.svg" } </script>
-<script type="application/ld+json"> { "@context": "http://schema.org", "@type": "NewsArticle", "mainEntityOfPage": { "@type": "WebPage", "@id": "https://bykvu.com/ru/bukvy/uchenye-nazvali-depressiju-prichinoj-22-opasnyh-zabolevanij/" }, "headline": "Ученые назвали депрессию причиной 22 опасных заболеваний", "image": [ "https://bykvu.com/wp-content/themes/bykvu/includes/images/noimage_large.jpg" ], "datePublished": "2019-09-09T00:29:09+02:00", "dateModified": "2019-09-09T00:29:09+02:00", "author": { "@type": "Person", "name": "Буквы" }, "publisher": { "@type": "Organization", "name": "Буквы", "logo": { "@type": "ImageObject", "url": "https://bykvu.com/wp-content/themes/bykvu/img/apple-icon-180x180.png" } }, "description": "Ученые австралийского центра точного здравоохранения при Университете Южной Австралии выяснили, что депрессия является причиной 22 различных заболеваний." } </script>
-<script type="application/ld+json"> { "@context": "https://schema.org", "@type": "BreadcrumbList", "itemListElement": [ { "@type": "ListItem", "position": 1, "item": { "@id": "https://bykvu.com/ru", "name": "Буквы" } }, { "@type": "ListItem", "position": 2, "item": { "@id": "https://bykvu.com/ru/category/bukvy/", "name": "Новости" } }, { "@type": "ListItem", "position": 3, "item": { "@id": "https://bykvu.com/ru/bukvy/uchenye-nazvali-depressiju-prichinoj-22-opasnyh-zabolevanij/", "name": "Ученые назвали депрессию причиной 22 опасных заболеваний" } } ] } </script>`)
+      const $ = cheerio.load(`<script type="application/ld+json"> { "@context": "http://schema.org", "@type": "Organization", "url": "https://bykvu.com/ru", "logo": "https://bykvu.com/wp-content/themes/bykvu/img/logo.svg" } </script>
+        <script type="application/ld+json"> { "@context": "http://schema.org", "@type": "NewsArticle", "mainEntityOfPage": { "@type": "WebPage", "@id": "https://bykvu.com/ru/bukvy/uchenye-nazvali-depressiju-prichinoj-22-opasnyh-zabolevanij/" }, "headline": "Ученые назвали депрессию причиной 22 опасных заболеваний", "image": [ "https://bykvu.com/wp-content/themes/bykvu/includes/images/noimage_large.jpg" ], "datePublished": "2019-09-09T00:29:09+02:00", "dateModified": "2019-09-09T00:29:09+02:00", "author": { "@type": "Person", "name": "Буквы" }, "publisher": { "@type": "Organization", "name": "Буквы", "logo": { "@type": "ImageObject", "url": "https://bykvu.com/wp-content/themes/bykvu/img/apple-icon-180x180.png" } }, "description": "Ученые австралийского центра точного здравоохранения при Университете Южной Австралии выяснили, что депрессия является причиной 22 различных заболеваний." } </script>
+        <script type="application/ld+json"> { "@context": "https://schema.org", "@type": "BreadcrumbList", "itemListElement": [ { "@type": "ListItem", "position": 1, "item": { "@id": "https://bykvu.com/ru", "name": "Буквы" } }, { "@type": "ListItem", "position": 2, "item": { "@id": "https://bykvu.com/ru/category/bukvy/", "name": "Новости" } }, { "@type": "ListItem", "position": 3, "item": { "@id": "https://bykvu.com/ru/bukvy/uchenye-nazvali-depressiju-prichinoj-22-opasnyh-zabolevanij/", "name": "Ученые назвали депрессию причиной 22 опасных заболеваний" } } ] } </script>`)
 
       const data = jsonld($)
       should(Object.keys(data).length).equal(3)
@@ -277,7 +276,8 @@ describe('metascraper-helpers', () => {
       <script type="application/ld+json">{ "offers": { "price": 119.99 }}</script>
       <script type="application/ld+json">{ "offers": { "price": "" }}</script>
       `)
-      should($jsonld('offers.price')($)).be.equal('119.99')
+      const value = $jsonld('offers.price')($)
+      should(value).be.equal(119.99)
     })
   })
 
@@ -350,7 +350,12 @@ describe('metascraper-helpers', () => {
     should(date('Jun 2018')).be.equal('2018-06-01T12:00:00.000Z')
     should(date(2010)).be.equal('2010-01-01T00:00:00.000Z')
     should(date(1594767608)).be.equal('2020-07-14T23:00:08.000Z')
+    should(date(1594767608 * 1000)).be.equal('2020-07-14T23:00:08.000Z')
+    should(date(1594767608 * 1000000)).be.equal('2020-07-14T23:00:08.000Z')
+    should(date(1594767608 * 1000000000)).be.equal('2020-07-14T23:00:08.000Z')
     should(date('11 juil. 2019')).be.equal(null)
+    const now = new Date()
+    should(date(now)).be.equal(now.toISOString())
   })
 })