+++ src/js/main.js
@@ -1,12 +1,10 @@
 /* global codecopy */
 
 window.$docsify = {
-  name: 'metascraper',
   repo: 'microlinkhq/metascraper',
-  maxLevel: 3,
+  maxLevel: 4,
   executeScript: true,
   auto2top: true,
-  ga: 'UA-108549225-3',
   noEmoji: true,
   plugins: [
     function (hook, vm) {