+++ packages/metascraper-title/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.23.1](https://nicedoc.io/microlinkhq/metascraper/packages/metascraper-title/compare/v5.23.0...v5.23.1) (2021-07-27)
+
+**Note:** Version bump only for package metascraper-title
+
+
+
+
+
+# [5.23.0](https://nicedoc.io/microlinkhq/metascraper/packages/metascraper-title/compare/v5.22.11...v5.23.0) (2021-07-26)
+
+**Note:** Version bump only for package metascraper-title
+
+
+
+
+
+## [5.22.11](https://nicedoc.io/microlinkhq/metascraper/packages/metascraper-title/compare/v5.22.10...v5.22.11) (2021-07-19)
+
+**Note:** Version bump only for package metascraper-title
+
+
+
+
+
+## [5.22.9](https://nicedoc.io/microlinkhq/metascraper/packages/metascraper-title/compare/v5.22.8...v5.22.9) (2021-07-11)
+
+**Note:** Version bump only for package metascraper-title
+
+
+
+
+
+## [5.22.8](https://nicedoc.io/microlinkhq/metascraper/packages/metascraper-title/compare/v5.22.7...v5.22.8) (2021-07-10)
+
+**Note:** Version bump only for package metascraper-title
+
+
+
+
+
+## [5.22.7](https://nicedoc.io/microlinkhq/metascraper/packages/metascraper-title/compare/v5.22.6...v5.22.7) (2021-07-08)
+
+**Note:** Version bump only for package metascraper-title
+
+
+
+
+
+## [5.22.5](https://nicedoc.io/microlinkhq/metascraper/packages/metascraper-title/compare/v5.22.4...v5.22.5) (2021-06-29)
+
+**Note:** Version bump only for package metascraper-title
+
+
+
+
+
 ## [5.22.4](https://nicedoc.io/microlinkhq/metascraper/packages/metascraper-title/compare/v5.22.3...v5.22.4) (2021-06-27)
 
 **Note:** Version bump only for package metascraper-title